Module aws_sdk_ssoadmin::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AttachManagedPolicyToPermissionSetError

ConflictException

Occurs when a conflict with a previous successful write is detected. This generally occurs

CreateAccountAssignmentError

CreateInstanceAccessControlAttributeConfigurationError

CreatePermissionSetError

DeleteAccountAssignmentError

DeleteInlinePolicyFromPermissionSetError

DeleteInstanceAccessControlAttributeConfigurationError

DeletePermissionSetError

DescribeAccountAssignmentCreationStatusError

DescribeAccountAssignmentDeletionStatusError

DescribeInstanceAccessControlAttributeConfigurationError

DescribePermissionSetError

DescribePermissionSetProvisioningStatusError

DetachManagedPolicyFromPermissionSetError

GetInlinePolicyForPermissionSetError

InternalServerException

The request processing has failed because of an unknown error, exception, or failure with

ListAccountAssignmentCreationStatusError

ListAccountAssignmentDeletionStatusError

ListAccountAssignmentsError

ListAccountsForProvisionedPermissionSetError

ListInstancesError

ListManagedPoliciesInPermissionSetError

ListPermissionSetProvisioningStatusError

ListPermissionSetsError

ListPermissionSetsProvisionedToAccountError

ListTagsForResourceError

ProvisionPermissionSetError

PutInlinePolicyToPermissionSetError

ResourceNotFoundException

Indicates that a requested resource is not found.

ServiceQuotaExceededException

Indicates that the principal has crossed the permitted number of resources that can be

TagResourceError

ThrottlingException

Indicates that the principal has crossed the throttling limits of the API

UntagResourceError

UpdateInstanceAccessControlAttributeConfigurationError

UpdatePermissionSetError

ValidationException

The request failed because it contains a syntax error.

Enums

AttachManagedPolicyToPermissionSetErrorKind

CreateAccountAssignmentErrorKind

CreateInstanceAccessControlAttributeConfigurationErrorKind

CreatePermissionSetErrorKind

DeleteAccountAssignmentErrorKind

DeleteInlinePolicyFromPermissionSetErrorKind

DeleteInstanceAccessControlAttributeConfigurationErrorKind

DeletePermissionSetErrorKind

DescribeAccountAssignmentCreationStatusErrorKind

DescribeAccountAssignmentDeletionStatusErrorKind

DescribeInstanceAccessControlAttributeConfigurationErrorKind

DescribePermissionSetErrorKind

DescribePermissionSetProvisioningStatusErrorKind

DetachManagedPolicyFromPermissionSetErrorKind

GetInlinePolicyForPermissionSetErrorKind

ListAccountAssignmentCreationStatusErrorKind

ListAccountAssignmentDeletionStatusErrorKind

ListAccountAssignmentsErrorKind

ListAccountsForProvisionedPermissionSetErrorKind

ListInstancesErrorKind

ListManagedPoliciesInPermissionSetErrorKind

ListPermissionSetProvisioningStatusErrorKind

ListPermissionSetsErrorKind

ListPermissionSetsProvisionedToAccountErrorKind

ListTagsForResourceErrorKind

ProvisionPermissionSetErrorKind

PutInlinePolicyToPermissionSetErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateInstanceAccessControlAttributeConfigurationErrorKind

UpdatePermissionSetErrorKind