Module aws_sdk_ssoadmin::model

Modules

access_control_attribute

See AccessControlAttribute

access_control_attribute_value

See AccessControlAttributeValue

account_assignment

See AccountAssignment

account_assignment_operation_status

See AccountAssignmentOperationStatus

account_assignment_operation_status_metadata

See AccountAssignmentOperationStatusMetadata

attached_managed_policy

See AttachedManagedPolicy

instance_access_control_attribute_configuration

See InstanceAccessControlAttributeConfiguration

instance_metadata

See InstanceMetadata

operation_status_filter

See OperationStatusFilter

permission_set

See PermissionSet

permission_set_provisioning_status

See PermissionSetProvisioningStatus

permission_set_provisioning_status_metadata

See PermissionSetProvisioningStatusMetadata

tag

See Tag

Structs

AccessControlAttribute

These are Amazon Web Services SSO identity store attributes that you can configure for use in

AccessControlAttributeValue

The value used for mapping a specified attribute to an identity source.

AccountAssignment

The assignment that indicates a principal's limited access to a specified Amazon Web Services account

AccountAssignmentOperationStatus

The status of the creation or deletion operation of an assignment that a principal needs

AccountAssignmentOperationStatusMetadata

Provides information about the AccountAssignment creation

AttachedManagedPolicy

A structure that stores the details of the IAM managed policy.

InstanceAccessControlAttributeConfiguration

Specifies the attributes to add to your attribute-based access control (ABAC)

InstanceMetadata

Provides information about the SSO instance.

OperationStatusFilter

Filters he operation status list based on the passed attribute value.

PermissionSet

An entity that contains IAM policies.

PermissionSetProvisioningStatus

A structure that is used to provide the status of the provisioning operation for a

PermissionSetProvisioningStatusMetadata

Provides information about the permission set provisioning status.

Tag

A set of key-value pairs that are used to manage the resource. Tags can only be applied to

Enums

InstanceAccessControlAttributeConfigurationStatus

PrincipalType

ProvisionTargetType

ProvisioningStatus

StatusValues

TargetType