Module aws_sdk_ssoadmin::operation

Structs

AttachManagedPolicyToPermissionSet

Attaches an IAM managed policy ARN to a permission set.

CreateAccountAssignment

Assigns access to a principal for a specified Amazon Web Services account using a specified permission

CreateInstanceAccessControlAttributeConfiguration

Enables the attributes-based access control (ABAC) feature for the specified Amazon Web Services SSO

CreatePermissionSet

Creates a permission set within a specified SSO instance.

DeleteAccountAssignment

Deletes a principal's access from a specified Amazon Web Services account using a specified permission

DeleteInlinePolicyFromPermissionSet

Deletes the inline policy from a specified permission set.

DeleteInstanceAccessControlAttributeConfiguration

Disables the attributes-based access control (ABAC) feature for the specified Amazon Web Services SSO

DeletePermissionSet

Deletes the specified permission set.

DescribeAccountAssignmentCreationStatus

Describes the status of the assignment creation request.

DescribeAccountAssignmentDeletionStatus

Describes the status of the assignment deletion request.

DescribeInstanceAccessControlAttributeConfiguration

Returns the list of Amazon Web Services SSO identity store attributes that have been configured to

DescribePermissionSet

Gets the details of the permission set.

DescribePermissionSetProvisioningStatus

Describes the status for the given permission set provisioning request.

DetachManagedPolicyFromPermissionSet

Detaches the attached IAM managed policy ARN from the specified permission set.

GetInlinePolicyForPermissionSet

Obtains the inline policy assigned to the permission set.

ListAccountAssignmentCreationStatus

Lists the status of the Amazon Web Services account assignment creation requests for a specified SSO

ListAccountAssignmentDeletionStatus

Lists the status of the Amazon Web Services account assignment deletion requests for a specified SSO

ListAccountAssignments

Lists the assignee of the specified Amazon Web Services account with the specified permission

ListAccountsForProvisionedPermissionSet

Lists all the Amazon Web Services accounts where the specified permission set is provisioned.

ListInstances

Lists the SSO instances that the caller has access to.

ListManagedPoliciesInPermissionSet

Lists the IAM managed policy that is attached to a specified permission set.

ListPermissionSetProvisioningStatus

Lists the status of the permission set provisioning requests for a specified SSO

ListPermissionSets

Lists the PermissionSets in an SSO instance.

ListPermissionSetsProvisionedToAccount

Lists all the permission sets that are provisioned to a specified Amazon Web Services account.

ListTagsForResource

Lists the tags that are attached to a specified resource.

ProvisionPermissionSet

The process by which a specified permission set is provisioned to the specified

PutInlinePolicyToPermissionSet

Attaches an IAM inline policy to a permission set.

TagResource

Associates a set of tags with a specified resource.

UntagResource

Disassociates a set of tags from a specified resource.

UpdateInstanceAccessControlAttributeConfiguration

Updates the Amazon Web Services SSO identity store attributes that you can use with the Amazon Web Services SSO

UpdatePermissionSet

Updates an existing permission set.